Better parsing of host XMLs from NCBI and deletion of corrupted/incorrectly-parsed files.

+++ db_config/database.py
@@ -286,7 +286,7 @@ class NucleotideVariant(_base):
 class VariantImpact(_base):
     __tablename__ = 'variant_impact'
 
-    variant_impact_id = Column(BigInteger, primary_key=True)
+    variant_impact_id = Column(Integer, primary_key=True)
     nucleotide_variant_id = Column(Integer, nullable=False)
 
     effect = Column(String)
